requirements.txt
asgiref==3.2.10
bson==0.5.8
dataclasses==0.6
Django==2.2.16
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djongo==1.3.0
Pillow==7.2.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
zipp==3.1.0
